python/requirements.txt
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
protobuf==3.19.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
